pyproject.toml
[tool.poetry]
name = "algorithmic-excercises"
version = "0.1.0"
description = ""
authors = ["percebus"]
license = "SEE LICENSE IN LICENSE.md"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "~3.12"
logger = "^1.4"
logging-config = "^1.0.4"
more-itertools = "^10.3.0"
pandas-stubs = "^2.2.3.241009"
pydantic-settings = "^2.3.4"
pyhamcrest = "^2.1.0"

[tool.poetry.group.dev.dependencies]
coverage = "^7.5.3"
flake8 = "^7.0.0"
mypy = "^1.13.0"
numpy = "^2.0.0"
pandas = "^2.2.2"
pre-commit = "^3.7.1"
pyclean = "^3.0.0"
pyhamcrest = "^2.1.0"
pylint = "^3.2.2"
pypyr = "^5.9.1"
pyright = "^1.1.345"
pytest = "^7.4.3"
pytest-bdd = "^7.0.1"
pytest-cov = "^5.0.0"
pytest-dotenv = "^0.5.2"
ruff = "^0.4.7"
setuptools = "^69.0.3"

[tool.pyright]
defineConstant = { DEBUG = true }
include = ["src"]
strict = ["src"]
# stubPath = "src/stubs" # TODO?
# ignore = ["src/oldstuff"] # NOTE what is the difference between exclude and ignore?
exclude = [
    "**/node_modules",
    "**/__pycache__",
]

[virtualenvs]
create = true
in-project = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
